added selection prompt
